Move js-only utils to `language-js` dir
