+++ src/language-js/needs-parens.js
@@ -2,12 +2,15 @@
 
 const assert = require("assert");
 
-const util = require("../common/util");
 const {
   getLeftSidePathName,
   hasFlowShorthandAnnotationComment,
   hasNakedLeftSide,
   hasNode,
+  isBitwiseOperator,
+  startsWithNoLookaheadToken,
+  shouldFlatten,
+  getPrecedence,
 } = require("./utils");
 
 function needsParens(path, options) {
@@ -139,15 +142,12 @@ function needsParens(path, options) {
     (parent.type === "ArrowFunctionExpression" &&
     parent.body === node &&
     node.type !== "SequenceExpression" && // these have parens added anyway
-      util.startsWithNoLookaheadToken(
+      startsWithNoLookaheadToken(
         node,
         /* forbidFunctionClassAndDoExpr */ false
       )) ||
     (parent.type === "ExpressionStatement" &&
-      util.startsWithNoLookaheadToken(
-        node,
-        /* forbidFunctionClassAndDoExpr */ true
-      ))
+      startsWithNoLookaheadToken(node, /* forbidFunctionClassAndDoExpr */ true))
   ) {
     return true;
   }
@@ -290,9 +290,9 @@ function needsParens(path, options) {
           }
 
           const po = parent.operator;
-          const pp = util.getPrecedence(po);
+          const pp = getPrecedence(po);
           const no = node.operator;
-          const np = util.getPrecedence(no);
+          const np = getPrecedence(no);
 
           if (pp > np) {
             return true;
@@ -303,7 +303,7 @@ function needsParens(path, options) {
             return true;
           }
 
-          if (pp === np && !util.shouldFlatten(po, no)) {
+          if (pp === np && !shouldFlatten(po, no)) {
             return true;
           }
 
@@ -313,7 +313,7 @@ function needsParens(path, options) {
 
           // Add parenthesis when working with bitwise operators
           // It's not strictly needed but helps with code understanding
-          if (util.isBitwiseOperator(po)) {
+          if (isBitwiseOperator(po)) {
             return true;
           }
